edited and tested add/remove moderator

+++ src/auth/models/user.js
@@ -269,6 +269,10 @@ const removeMod = async (mobile) => {
         let safeValues = [mobile];
         let result = await client.query(SQL, safeValues);
 
+        if(!result.rows[0]){
+            return -1;
+        }
+
         let userId = result.rows[0].id;
         SQL = `DELETE FROM MODERATOR WHERE user_id=$1;`;
         safeValues = [userId];
